Refactored conditions of changing state in State Design Pattern.

+++ StateAndDecoratorDesignPatterns/SignalModelContext.cs
@@ -71,7 +71,7 @@ namespace StateAndDecoratorDesignPattern
             }
         }
 
-        public void SetSignalValue(int signalValue)
+        public void SetSignalValue(double signalValue)
         {
             CurrentState.SetSignalValue(signalValue);
         }
